[Feat] spark support.
