+++ streamx-spark/streamx-spark-connector/pom.xml
@@ -1,0 +1,84 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <parent>
+        <groupId>com.streamxhub.streamx</groupId>
+        <artifactId>streamx-spark</artifactId>
+        <version>1.2.4</version>
+    </parent>
+
+    <artifactId>streamx-spark-connector_${scala.binary.version}</artifactId>
+
+    <name>StreamX : Spark Connector</name>
+    <packaging>pom</packaging>
+
+    <properties>
+        <vertx.version>4.2.3</vertx.version>
+        <clickhouse.version>0.3.2</clickhouse.version>
+        <async.client.version>2.12.3</async.client.version>
+        <influxdb.version>2.17</influxdb.version>
+    </properties>
+
+    <modules>
+        <module>streamx-spark-connector-base</module>
+        <module>streamx-spark-connector-kafka</module>
+    </modules>
+
+    <dependencyManagement>
+
+        <dependencies>
+
+            <dependency>
+                <groupId>com.streamxhub.streamx</groupId>
+                <artifactId>streamx-common_${scala.binary.version}</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+
+            <dependency>
+                <groupId>com.streamxhub.streamx</groupId>
+                <artifactId>streamx-spark-core_${scala.binary.version}</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+
+            <dependency>
+                <groupId>com.streamxhub.streamx</groupId>
+                <artifactId>streamx-spark-connector-base_${scala.binary.version}</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+
+        </dependencies>
+
+    </dependencyManagement>
+
+    <profiles>
+        <profile>
+            <id>connector_2.12</id>
+            <activation>
+                <activeByDefault>true</activeByDefault>
+                <property>
+                    <name>scala.binary.version</name>
+                    <value>2.12</value>
+                </property>
+            </activation>
+            <modules>
+
+            </modules>
+        </profile>
+
+        <profile>
+            <id>connector_2.11</id>
+            <activation>
+                <property>
+                    <name>scala.binary.version</name>
+                    <value>2.11</value>
+                </property>
+            </activation>
+            <modules>
+                <!-- redis only support scala 2.11 -->
+<!--                <module>streamx-spark-connector-redis</module>-->
+            </modules>
+        </profile>
+    </profiles>
+
+</project>
