+++ streamx-spark/streamx-spark-connector/streamx-spark-connector-base/pom.xml
@@ -1,0 +1,31 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+
+    <parent>
+        <groupId>com.streamxhub.streamx</groupId>
+        <artifactId>streamx-spark-connector_${scala.binary.version}</artifactId>
+        <version>1.2.4</version>
+        <relativePath>../pom.xml</relativePath>
+    </parent>
+
+    <artifactId>streamx-spark-connector-base_${scala.binary.version}</artifactId>
+
+    <name>StreamX : Spark Connector Base</name>
+
+    <dependencies>
+
+        <dependency>
+            <groupId>com.streamxhub.streamx</groupId>
+            <artifactId>streamx-common_${scala.binary.version}</artifactId>
+        </dependency>
+
+        <dependency>
+            <groupId>com.streamxhub.streamx</groupId>
+            <artifactId>streamx-spark-core_${scala.binary.version}</artifactId>
+        </dependency>
+
+    </dependencies>
+
+</project>
