+++ streamx-spark/streamx-spark-connector/streamx-spark-connector-kafka/pom.xml
@@ -1,0 +1,30 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <parent>
+        <groupId>com.streamxhub.streamx</groupId>
+        <artifactId>streamx-spark-connector_${scala.binary.version}</artifactId>
+        <version>1.2.4</version>
+    </parent>
+
+    <artifactId>streamx-spark-connector-kafka_${scala.binary.version}</artifactId>
+
+    <name>StreamX : Spark Connector Kafka</name>
+
+    <dependencies>
+
+        <dependency>
+            <groupId>com.streamxhub.streamx</groupId>
+            <artifactId>streamx-spark-connector-base_${scala.binary.version}</artifactId>
+        </dependency>
+
+        <dependency>
+            <groupId>org.apache.kafka</groupId>
+            <artifactId>kafka_${scala.binary.version}</artifactId>
+            <version>0.10.2.2</version>
+        </dependency>
+
+    </dependencies>
+
+</project>
