--- streamx-spark/streamx-spark-core/src/main/scala/com/streamxhub/streamx/spark/core/support/kafka/writer/DStreamKafkaWriter.scala
@@ -1,47 +1,0 @@
-/*
- * Copyright (c) 2019 The StreamX Project
- *
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.streamxhub.streamx.spark.core.support.kafka.writer
-
-import org.apache.kafka.clients.producer.ProducerRecord
-import org.apache.spark.streaming.dstream.DStream
-
-import java.util.Properties
-import scala.annotation.meta.param
-import scala.reflect.ClassTag
-
-class DStreamKafkaWriter[T: ClassTag](@(transient@param) dstream: DStream[T]) extends KafkaWriter[T] {
-
-  /**
-    *
-    * @param producerConfig The configuration that can be used to connect to Kafka
-    * @param serializerFunc The function to convert the data from the stream into Kafka
-    *                       [[ProducerRecord]]s.
-    * @tparam K The type of the key
-    * @tparam V The type of the value
-    *
-    */
-  override def writeToKafka[K, V](producerConfig: Properties,
-                                  serializerFunc: T => ProducerRecord[K, V]): Unit = {
-    dstream.foreachRDD { rdd =>
-      val rddWriter = new RDDKafkaWriter[T](rdd)
-      rddWriter.writeToKafka(producerConfig, serializerFunc)
-    }
-  }
-}
