+++ streamx-spark/streamx-spark-test/assembly/conf/prod/test.properties
@@ -95,7 +95,7 @@ spark.dynamicAllocation.sustainedSchedulerBacklogTimeout=1s
 ######################################################
 spark.source.kafka.consume.topics=bigdata
 spark.source.kafka.consume.group.id=test
-spark.source.kafka.consume.bootstrap.servers=localhost:9091,localhost:9092,localhost:9093
+spark.source.kafka.consume.bootstrap.servers=127.0.0.1:9094,127.0.0.1:9092,127.0.0.1:9093
 spark.source.kafka.consume.key.deserializer=org.apache.kafka.common.serialization.StringDeserializer
 spark.source.kafka.consume.value.deserializer=org.apache.kafka.common.serialization.StringDeserializer
 spark.source.kafka.consume.fetch.max.wait.ms=3000
