+++ streamx-spark/streamx-spark-test/src/main/resources/logback.xml
@@ -113,7 +113,7 @@
     </appender>
 
     <!-- Logger 根目录 -->
-    <root level="DEBUG">
+    <root level="INFO">
         <appender-ref ref="STDOUT"/>
         <appender-ref ref="DEBUG"/>
         <appender-ref ref="ERROR"/>
